# docs/swagger/index.yml
openapi: 3.0.3
info:
  title: online-shop-express
  version: 1.0.0
  contact: {}
servers:
  - url: 127.0.0.1
paths:
...
  /api/v1/wallet:
    get:
      tags:
        - wallet-service
      summary: read-wallet
      description: read-wallet
      operationId: readWallet
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
  /api/v1/wallet/deposit:
    patch:
      tags:
        - wallet-service
      summary: deposit
      description: deposit
      operationId: deposit
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                amount:
                  type: number
                  example: 250000
            examples:
              deposit:
                value:
                  amount: 250000
      responses:
        '200':
          description: ''
      security:
        - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
...